Almost 9ms, actually use FMA instructions
Refactor BVH and HittableList for improved performance

- Replaced `BVHNode` struct with separate left and right child arrays for better memory access and performance.
- Updated `Init`, `AddNode`, and `Build` functions to accommodate the new structure.
- Modified traversal logic to utilize the new child arrays.
- Changed `HittableList` to use `glm::vec4` for centers, enhancing data storage and alignment.
- Refactored `Add` and `IntersectPrimitive` functions for improved handling of centers and normals.
- Optimized `Material` struct with `__restrict__` qualifiers and refactored `Scatter` for clarity and efficiency.
- Added `RandomVec2` function for generating 2D random vectors.
- Updated loop counters in `RayColor` and `RenderParams` to use `uint16_t` for reduced memory usage.
- Included `std::numbers` in `pch.h` for better numerical constants.
- Enhanced `Reflectance` function with a more accurate Schlick's approximation.
- Introduced a new branchless `RefractBranchless` function for improved refraction performance.

Refactor BVH structure and optimize memory usage

- Removed `BVHNode` struct; introduced `m_LeftNodes` and `m_RightNodes` for child indices to enhance memory access patterns.
- Updated `Init` and `AddNode` functions to allocate and use new arrays.
- Modified traversal logic in BVH to accommodate new structure.
- Improved clarity in `CUDARender.cu` with explicit casting for `pixelIndex`.
- Adjusted member initialization order in `Camera` constructor.
- Enhanced `CudaGLSurface` with additional resource management variables.
- Added `__restrict__` qualifiers in `HittableList.h` and `Material.h` for performance optimization.
- Simplified `Render` function in `Main.cpp` by removing unnecessary parameters.
- Introduced `RandomVec2` function in `Random.h` for 2D random vector generation.
- Changed loop counters in `RayColor` to `uint16_t` for reduced memory usage.
- Updated `Renderer.h` to use `uint16_t` for `m_SamplesPerPixel` and `m_MaxDepth`.
- Improved reflection and refraction calculations in `Vec3.h` with a branchless refract function.
- Added `Vec2` and `Vec3` type definitions in `pch.h` for consistency.

+++ src/Renderer.h
@@ -3,6 +3,7 @@
 #include "Camera.h"
 #include <chrono>
 
+#include "AABB.h"
 #include "CPU_GPU.h"
 #include "SFML/System/Vector2.hpp"
 
@@ -37,8 +38,8 @@ struct RenderParams
 	CameraPOD			Camera {};
 	float4				ResolutionInfo {}; // x: x Pixel Size, y: y Pixel Size, z: width, w: height
 	cudaSurfaceObject_t Image {};
-	uint32_t			m_SamplesPerPixel {};
-	uint32_t			m_MaxDepth {};
+	uint16_t			m_SamplesPerPixel {};
+	uint16_t			m_MaxDepth {};
 	float				m_ColorMul {};
 };
